requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.5
eth-account==0.13.7
python-dotenv==1.0.1
pydantic==2.9.2
PyJWT==2.9.0
bcrypt==5.0.0
cryptography>=42.0.0
slowapi==0.1.9
